Add files via upload
Here I provide some visualization using seaborn data visualization library, such as : countplot, barplot, jointplot, and pairplot from 2 dataset.

First dataset is iris, which is collected from sklearn.dataset (module in scikit-learn library).  sklearn.datasets provides various functions and classes for loading and generating datasets that are commonly used for machine learning tasks. The module offers a collection of pre-existing datasets that are included with scikit-learn, such as the Iris dataset, the Boston Housing dataset, the MNIST (Modified National Institute of Standards and Technology-->collect handwritten digits) dataset, and many others. 

The Iris dataset consists of 150 samples, where each sample represents an iris flower. For each sample, there are four features or attributes that describe the flower:
Sepal Length: The length of the sepal (in centimeters).
Sepal Width: The width of the sepal (in centimeters).
Petal Length: The length of the petal (in centimeters).
Petal Width: The width of the petal (in centimeters).
Additionally, each sample is labeled with the species of the iris flower it represents. There are three possible classes or species in the dataset:
Setosa, Versicolor, Virginica. The goal of using the Iris dataset is typically to develop and evaluate machine learning algorithms for classification tasks. Given the four features of an iris flower, the task is to predict its corresponding species.

Second dataset is udemy course which provided by Kaggle. Here the link : https://www.kaggle.com/datasets/andrewmvd/udemy-courses 
This dataset contains 3.682 records of courses from 4 subjects (Business Finance, Graphic Design, Musical Instruments and Web Design) taken from Udemy.
*Udemy is a massive online open course (MOOC) platform that offers both free and paid courses. Anybody can create a course, a business model by which allowed Udemy to have hundreds of thousands of courses